Refactor PropertyType
