+++ src/main/scala/com/cognite/sdk/scala/v1/fdm/common/properties/PrimitivePropType.scala
@@ -1,7 +1,7 @@
 // Copyright 2020 Cognite AS
 // SPDX-License-Identifier: Apache-2.0
 
-package com.cognite.sdk.scala.v1.fdm.containers
+package com.cognite.sdk.scala.v1.fdm.common.properties
 
 import io.circe.{Decoder, Encoder}
